Fix various breakages in Windows process handling that creeped in because we weren't actually running the test:
- Call TerminateProcess() only if the process handle is still open
- Update the tests so that they expect a return value of 0 when reading a stream from a non-existent process

Fixes #1538 .

--
Reviewed-on: https://bazel-review.googlesource.com/4100
MOS_MIGRATED_REVID=127935621

+++ src/main/native/windows_processes.cc
@@ -450,7 +450,7 @@ Java_com_google_devtools_build_lib_windows_WindowsProcesses_nativeTerminate(
       process->error_ = GetLastErrorString("TerminateJobObject()");
       return JNI_FALSE;
     }
-  } else {
+  } else if (process->process_ != INVALID_HANDLE_VALUE) {
     if (!TerminateProcess(process->process_, 1)) {
       process->error_ = GetLastErrorString("TerminateProcess()");
       return JNI_FALSE;
